pyproject.toml
[tool.poetry]
name = "collaborative-filtering"
version = "0.1.0"
description = ""
authors = ["charleslow <charleslow88@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
lightning = "^2.5.0.post0"
mlflow-skinny = "^2.19.0"
numpy = "^1.26.4"
pyarrow = "^18.1.0"
sentence-transformers = "^3.3.1"
torch = "^2.4.0"
transformers = "^4.47.1"
xxhash = "^3.5.0"
numpydantic = "^1.6.6"
tensorboard = "^2.18.0"
pandas = "^2.2.3"
scipy = "^1.15.1"
recommenders = "^1.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
